pyproject.toml
[tool.poetry]
name = "rfdocsindexer"
version = "0.1.0"
description = ""
authors = ["Vincent Maire <maire.vincent31@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
indexrfdocs = "rfdocsindexer.cli:cli"

[tool.poetry.dependencies]
python = "^3.6.2"
Jinja2 = "^3.0.1"
robotframework = "^4"
click = "^8.0.1"
pydantic = "^1.8.2"
toml = "^0.10.2"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
black = "^21.7b0"
mypy = "^0.910"
flake8 = "^3.9.2"
types-toml = "^0.1.4"
tox = "^3.24.1"
tox-poetry-installer = {extras = ["poetry"], version = "^0.8.1"}
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
